domparsing/serializing_html_doctypevariation.html

<!DoCtYpE hTmL  SyStEm     " external.dtd " [  <!ENTITY   blah    "val-value "&gt; ]>
<html>
 <head>
  <title>DOCTYPE Serialization in HTML - variation 1</title>
  <script src="../resources/testharness.js"></script>
  <script src="../resources/testharnessreport.js"></script>
 </head>
 <body>
  <div id="log"></div>
  <script>
  test(function(){
    assert_equals(new XMLSerializer().serializeToString(document.doctype), 
                  "<!DOCTYPE html SYSTEM \" external.dtd \">");
  }, "Parsed DOCTYPE serialized correctly in HTML");
  </script>
 </body>
</html>
